fix(analytics): capture title and name properties
fix(analytics): capture title and name properties
