+++ middleware/trackPage.global.ts
@@ -2,6 +2,6 @@ export default defineNuxtRouteMiddleware((to) => {
   // Analytics must be executed only in client side
   if (process.client) {
     const { trackPage } = useSegment();
-    trackPage(window, to.fullPath);
+    trackPage(to.meta.routeName, to.meta.pageTitle);
   }
 });
