+++ types/modules.d.ts
@@ -1,0 +1,1 @@
+declare module "frontmatter-markdown-loader/mode";
